# docs/openapi.yaml
openapi: 3.1.0
info:
  title: A.T.O.M Planning API
  version: 1.0.0
  description: Deterministic, local-first RF planning endpoints. Results are planning estimates, not UE, drive-test, or PHY measurements.
servers:
  - url: http://localhost:8080
paths:
...
  /api/core/topology:
    get:
      summary: Xn-C, Xn-U, N2, and N3 topology
      parameters:
        - name: cluster_tower_ids
          in: query
          description: Comma-separated tower IDs; at most 64 entries and 128 UTF-8 bytes per ID.
          schema: { type: string }
        - { name: network_tech, in: query, schema: { type: string, const: 5g } }
      responses: { "200": { description: gNB nodes, interface edges, and route decisions } }
  /api/core/sessions:
    get:
      summary: Optional Core Lab sessions
      parameters:
        - name: cluster_tower_ids
          in: query
          description: Comma-separated tower IDs; at most 64 entries and 128 UTF-8 bytes per ID.
          schema: { type: string }
        - { name: network_tech, in: query, schema: { type: string, const: 5g } }
      responses: { "200": { description: Selected-cell session state } }
...
  /api/core/scenario:
    post:
      summary: Activate a Core Lab scenario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scenario]
              properties:
                scenario: { type: string, enum: [normal, registration_storm, udm_outage, ausf_auth_failure, pcf_policy_degraded, upf_degraded, xn_degraded, xn_unavailable] }
                cluster_tower_ids:
                  type: array
                  maxItems: 64
                  description: Tower IDs are trimmed, deduplicated, and limited to 128 UTF-8 bytes each.
                  items: { type: string }
                network_tech: { type: string, const: 5g }
      responses:
        "200": { description: Scenario accepted }
        "400": { $ref: "#/components/responses/BadRequest" }
components:
  responses:
    BadRequest: { description: Invalid request, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties: { error: { type: string } }
...